Migrate creating the driver service to the new architecture.
